Remove tsparse element from the bin after it's no longer needed
tsparse is only needed to obtain PCR's that are passed to tsindexer.
Once this is done we don't need to parse the stream. By removing
tsparse it's possible to lower the CPU usage by a few percen
(relative to the other elements in the tsshifterbin).

To make removal of element in a live pipeline possible, a fix had to
be applied in tsshifter, because it was dropping caps event which
was upsetting upstream elements (non-negotiated error).