Reduce timeshifter memory usage by removing slots.
Timeshifter uses tscache component to store buffers on the disk.
Tscache would divide the file (to which data is buffered) into small
chunks and every chunk would need a "slot" structure in memory for
housekeeping. For large files this can be many megabytes of memory
that is rarely used. Because data is accessed linearly there's no
need to keep such an expensive structure around. A simpler approach
with write-head and read-head pointers could be used to eliminate
the need for "slots".