Decrease timeshifter memory usage by replacing memindex
Timeshifter used Byzantine memindex class that was extremaly expensive
in terms of memory required for it's storage. This had an unpleasant potential
to use all the memory it could get, especially if used for many hours or days.
This is not an unlikely scenarion in set-top-box world hence there was a need
to replace that lavish complexity with something simpler. Ideally with fixed
memory usage.

This was achieved by introduction of simpleindex. It's a dead simple thing
that stored time:offset pairs in a file and allows to look them up.

With that change timeshifter's memory usage should be in the range of
several hundreds of kilobytes.